CRUD_Modifie

+++ flask_base/src/services/songs.py
@@ -62,7 +62,6 @@ def isUuidValid(id):
         # La conversion en UUID a échoué
         return False
 
-
 def addRatingSong(idSong,rating_add):
 
     if not isUuidValid(idSong):
@@ -83,23 +82,77 @@ def addRatingSong(idSong,rating_add):
   #J'attend l' id rating get song pour pouvoir renvoyer l'objet  
 
     return response_ratings.json(), response_ratings.status_code
-def create_song(song_data):
-    song_schema = CreateSongSchema().loads(json.dumps(song_data), unknown=EXCLUDE)
-    response = requests.request(method="POST", url=songs_url, json=song_schema)
 
-    if response.status_code != 201:
-        return response.json(), response.status_code
 
-    # Ajouter la chanson dans la base de données
+
+def isSongIdValid(id):
+    verif = requests.request(method="GET", url=songs_url+(id))
+    if verif.status_code == 200:
+        return True
+    else:
+        return False
+
+
+
+def isUuidValid(id):
     try:
-        song_model = SongSchema().load(response.json())
-        songs_repository.create_song(song_model)
-    except Exception:
+        uuid_obj = uuid.UUID(id)
+        return uuid_obj.version == 4
+    except ValueError:
+        # La conversion en UUID a échoué
+        return False
+
+
+
+def isUserIdValid(id):
+    verif = requests.request(method="GET", url=songs_url+id)
+    if verif.json()["idUser"] == current_user.id:
+        return True
+    else:
+        return False
+
+
+def get_song(id):
+
+    if not isUuidValid(id):
+        raise UnprocessableEntity
+    if not isSongIdValid(id):
+        raise NotFound
+
+    response = requests.get(songs_url + id)
+
+    if response.status_code == 500:
+        raise SomethingWentWrong  
+    return response.json(), response.status_code
+
+    if not isUuidValid(idSong):
+        raise UnprocessableEntity
+    if not isSongIdValid(idSong):
+        raise NotFound
+
+    rating_add["idSong"] = idSong
+    rating_add["idUser"] = current_user.id
+
+    UpdateRating_schema = UpdateRatingSchema().loads(json.dumps(rating_add), unknown=EXCLUDE)
+
+    response_ratings = requests.request(method="POST", url=ratings_url,json=UpdateRating_schema)
+
+    if response_ratings.status_code == 500:
         raise SomethingWentWrong
 
   #J'attend l' id rating get song pour pouvoir renvoyer l'objet  
 
     return response_ratings.json(), response_ratings.status_code
+def create_song(song_data):
+    
+    song_schema = CreateSongSchema().loads(json.dumps(song_data), unknown=EXCLUDE)
+    response = requests.request(method="POST", url=songs_url, json=song_schema)
+    
+    if response.status_code == 500:
+        raise SomethingWentWrong
+
+    return "", response.status_code
+
 
 def deleteRatingtoSong(idSong,idRating):
 
@@ -141,21 +194,23 @@ def setRatingtoSong(idSong,idRating,rating_upt):
         raise SomethingWentWrong
 
     return response_ratings.json(), response_ratings.status_code
+def get_song_from_db(song_id):
+    return songs_repository.get_song(song_id)
 
-def update_song(id, updated_data):
-    song_schema = UpdateSongSchema().loads(json.dumps(updated_data), unknown=EXCLUDE)
-    response = requests.request(method="PUT", url=songs_url + id, json=song_schema)
 
-    if response.status_code != 200:
-        return response.json(), response.status_code
+def song_exists(song_id):
+    return get_song_from_db(song_id) is not None
 
-    # Mettre à jour la chanson dans la base de données
-    try:
-        updated_song_model = SongSchema().load(response.json())
-        songs_repository.update_song(updated_song_model)
-    except Exception:
+def update_song(id, song_update):
+    if not isUuidValid(id):
+        raise UnprocessableEntity
+    if not isSongIdValid(id):
+        raise NotFound
+
+    response = requests.request(method="PUT", url=songs_url+id,json=song_update)
+
+    if response.status_code == 500:
         raise SomethingWentWrong
-
     return response.json(), response.status_code
 
 
@@ -166,12 +221,12 @@ def get_songs():
 
 
 def delete_song(id):
+    if not isSongIdValid(id):
+        raise NotFound
+   
     response=requests.request(method="DELETE", url=songs_url + id)
 
-    try:
-        songs_repository.delete_song(id)
-    except Exception:
-        error = NotFoundSchema().loads("{}")
-        return error, error.get("code")
+    if response.status_code == 500:
+        raise SomethingWentWrong
 
-    return response.json(), 200
+    return "", response.status_code
